Add an example of JSON Hyper-Schema document in Overview
This example is resurrected from jsonschema-hyperschema.xml prior to the
rewrite. The only change is that "authorId" is moved from "required" to
"hrefRequired". The idea of this section is to provide an easily
understable example as soon as possible in the document while not
requiring too much explanation. It appears in its own section so that
readers only interested in the implementation requirements could easily
skip it.